Entry.HW 클래스화
- PR #1324, #1333 적용
- feature/module 적용
