+++ src/entry.js
@@ -32,8 +32,7 @@ require('./class/playground');
 require('./class/popup');
 require('./class/popup_helper');
 require('./class/popup_list');
-// require('./class/hw');
-require('./class/hw2');
+require('./class/hw');
 require('./class/moduleManager');
 require('./class/hw_monitor');
 require('./class/project');
